+++ src/fncs/pages/index.html
@@ -148,6 +148,7 @@
 						<div class="cont_workspace_container">
 							<div class="cont_sitemap_parent"></div>
 							<div class="cont_sitemap_broslist"></div>
+							<div class="cont_sitemap_search"></div>
 						</div><!-- /.cont_workspace_container -->
 					</div>
 				</div>
