+++ src/fncs/pages/index_files/style.css.scss
@@ -106,3 +106,12 @@
 		overflow: hidden;
 	}
 }
+
+.cont_sitemap_search{
+	position: absolute;
+	top: 0; left: 0;
+	width: 100%;
+	height: 100%;
+	overflow: auto;
+	background-color: #fff;
+}
